Extract pairing logic and add tests
